chore: refactor advisory systems table
